updated game setup
