--- lib/turn2.rb
@@ -1,12 +1,0 @@
-require 'colorize'
-
-class Turn
-
-    def initialize 
-        
-    end
-
-    def take_turn
-
-    end
-end
